finished seperating all endpoints for managing receipts into the manageReceipts module
